integrated json-db

+++ init.sh
@@ -44,16 +44,43 @@ apt-get install -y php5-memcache php5-memcached
 echo ""
 echo "================ INSTALL client-ident-proxy =================="
 
-wget https://github.com/huangqiheng/client-ident-proxy/archive/master.tar.gz
-tar xzvf master.tar.gz
-mv client-ident-proxy-master client-ident-proxy
-rm master.tar.gz
+wwwroot=$current_path
+
+if [ ! -f $current_path/run ]; then
+	wget https://github.com/huangqiheng/client-ident-proxy/archive/master.tar.gz
+	tar xzvf master.tar.gz
+	mv client-ident-proxy-master client-ident-proxy
+	rm master.tar.gz
+
+	wwwroot=$current_path/client-ident-proxy
+fi
+
+echo ""
+echo "================ INSTALL jsondb =================="
+
+jsondb=$wwwroot/json-db
+
+if [ ! -f $jsondb ]; then
+	git clone https://github.com/huangqiheng/json-db.git
+
+	cd $jsondb
+
+	mkdir -p ./databases/jsondb.gw/gateway/proxy
+	cp $wwwroot/schema-db.json $jsondb/databases/jsondb.gw/gateway/schema.json
+	cp $wwwroot/schema-tb.json $jsondb/databases/jsondb.gw/gateway/proxy/schema.json
+
+	chown www-data:www-data ./databases -R
+	chown www-data:www-data ./cache -R
+	chown www-data:www-data ./admin/uploads -R
+	chown www-data:www-data ./service/queue/cache -R
+fi
 
 echo ""
 echo "================ START web application firwall =================="
 
-wwwroot=$current_path/client-ident-proxy
-sed -i "s|root.*client-ident-proxy;$|root $wwwroot;|g" $wwwroot/nginx.conf
+nginx_conf=$wwwroot/nginx.conf
+sed -i "s|root.*client-ident-proxy;$|root $wwwroot;|g" $nginx_conf
+sed -i "s|root.*json-db;$|root $jsondb;|g" $nginx_conf
 
 cd $wwwroot
 sh run
